Convert the add transaction and item methods over to the _JS class
